Fixed narrowing convertions

+++ main.cpp
@@ -196,12 +196,12 @@ GPIOController::GPIOController()
     gpio = new GPIO[GPIO_COUNT];
     for (uint8_t i = 0; i < GPIO_COUNT; i++) {
         gpio[i] = {
-            (uint32_t)i + 2,
+            (uint8_t)(i + 2),
             (uint32_t *)getPeripheral(GPIO_SET0_OFFSET), 
             (uint32_t *)getPeripheral(GPIO_CLR0_OFFSET), 
             (uint32_t *)getPeripheral(GPIO_LEVEL0_OFFSET),
-            (uint32_t *)getPeripheral(GPIO_FSEL_BASE_OFFSET + (((uint32_t)i + 2) * 3) / 30 * sizeof(uint32_t)), 
-            (((uint32_t)i + 2) * 3) % 30, 
+            (uint32_t *)getPeripheral(GPIO_FSEL_BASE_OFFSET + ((uint32_t)(i + 2) * 3) / 30 * sizeof(uint32_t)), 
+            ((uint32_t)(i + 2) * 3) % 30, 
             GPIO_MODE_UNKNOWN, 
             20.0, 
             1.0
